+++ airbyte-integrations/connectors/source-linkedin-ads/source_linkedin_ads/streams.py
@@ -5,7 +5,7 @@
 
 import logging
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Iterable, Mapping, MutableMapping, Optional, Union
+from typing import Any, Dict, Iterable, Mapping, MutableMapping, Optional
 from urllib.parse import urlencode
 
 import pendulum
