+++ airbyte-integrations/connectors/source-linkedin-ads/unit_tests/output_slices.json
@@ -1,0 +1,118 @@
+[
+  [
+    {
+      "campaign_id": 123,
+      "dateRange": {
+        "end.day": 31,
+        "end.month": 1,
+        "end.year": 2021,
+        "start.day": 1,
+        "start.month": 1,
+        "start.year": 2021
+      },
+      "fields": "actionClicks,adUnitClicks,approximateUniqueImpressions,cardClicks,cardImpressions,clicks,commentLikes,comments,companyPageClicks,conversionValueInLocalCurrency,costInLocalCurrency,costInUsd,dateRange,documentCompletions,documentFirstQuartileCompletions,documentMidpointCompletions,documentThirdQuartileCompletions,downloadClicks,externalWebsiteConversions"
+    }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 31,
+      "end.month": 1,
+      "end.year": 2021,
+      "start.day": 1,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "externalWebsitePostClickConversions,externalWebsitePostViewConversions,follows,fullScreenPlays,impressions,jobApplications,jobApplyClicks,landingPageClicks,leadGenerationMailContactInfoShares,leadGenerationMailInterestedClicks,likes,oneClickLeadFormOpens,oneClickLeads,opens,otherEngagements,pivotValues,postClickJobApplications,postClickJobApplyClicks,postClickRegistrations,dateRange"
+  }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 31,
+      "end.month": 1,
+      "end.year": 2021,
+      "start.day": 1,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "postViewJobApplications,postViewJobApplyClicks,postViewRegistrations,reactions,registrations,sends,shares,talentLeads,textUrlClicks,totalEngagements,validWorkEmailLeads,videoCompletions,videoFirstQuartileCompletions,videoMidpointCompletions,videoStarts,videoThirdQuartileCompletions,videoViews,viralCardClicks,viralCardImpressions,dateRange"
+  }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 31,
+      "end.month": 1,
+      "end.year": 2021,
+      "start.day": 1,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "viralClicks,viralCommentLikes,viralComments,viralCompanyPageClicks,viralDocumentCompletions,viralDocumentFirstQuartileCompletions,viralDocumentMidpointCompletions,viralDocumentThirdQuartileCompletions,viralDownloadClicks,viralExternalWebsiteConversions,viralExternalWebsitePostClickConversions,viralExternalWebsitePostViewConversions,viralFollows,viralFullScreenPlays,viralImpressions,viralJobApplications,viralJobApplyClicks,viralLandingPageClicks,viralLikes,dateRange"
+  }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 31,
+      "end.month": 1,
+      "end.year": 2021,
+      "start.day": 1,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "viralOneClickLeadFormOpens,viralOneClickLeads,viralOtherEngagements,viralPostClickJobApplications,viralPostClickJobApplyClicks,viralPostClickRegistrations,viralPostViewJobApplications,viralPostViewJobApplyClicks,viralPostViewRegistrations,viralReactions,viralRegistrations,viralShares,viralTotalEngagements,viralVideoCompletions,viralVideoFirstQuartileCompletions,viralVideoMidpointCompletions,viralVideoStarts,viralVideoThirdQuartileCompletions,viralVideoViews,dateRange"
+  }
+  ],
+  [
+    {
+      "campaign_id": 123,
+      "dateRange": {
+        "end.day": 2,
+        "end.month": 3,
+        "end.year": 2021,
+        "start.day": 31,
+        "start.month": 1,
+        "start.year": 2021
+      },
+      "fields": "actionClicks,adUnitClicks,approximateUniqueImpressions,cardClicks,cardImpressions,clicks,commentLikes,comments,companyPageClicks,conversionValueInLocalCurrency,costInLocalCurrency,costInUsd,dateRange,documentCompletions,documentFirstQuartileCompletions,documentMidpointCompletions,documentThirdQuartileCompletions,downloadClicks,externalWebsiteConversions"
+    }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 2,
+      "end.month": 3,
+      "end.year": 2021,
+      "start.day": 31,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "externalWebsitePostClickConversions,externalWebsitePostViewConversions,follows,fullScreenPlays,impressions,jobApplications,jobApplyClicks,landingPageClicks,leadGenerationMailContactInfoShares,leadGenerationMailInterestedClicks,likes,oneClickLeadFormOpens,oneClickLeads,opens,otherEngagements,pivotValues,postClickJobApplications,postClickJobApplyClicks,postClickRegistrations,dateRange"
+  }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 2,
+      "end.month": 3,
+      "end.year": 2021,
+      "start.day": 31,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "postViewJobApplications,postViewJobApplyClicks,postViewRegistrations,reactions,registrations,sends,shares,talentLeads,textUrlClicks,totalEngagements,validWorkEmailLeads,videoCompletions,videoFirstQuartileCompletions,videoMidpointCompletions,videoStarts,videoThirdQuartileCompletions,videoViews,viralCardClicks,viralCardImpressions,dateRange"
+  }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 2,
+      "end.month": 3,
+      "end.year": 2021,
+      "start.day": 31,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "viralClicks,viralCommentLikes,viralComments,viralCompanyPageClicks,viralDocumentCompletions,viralDocumentFirstQuartileCompletions,viralDocumentMidpointCompletions,viralDocumentThirdQuartileCompletions,viralDownloadClicks,viralExternalWebsiteConversions,viralExternalWebsitePostClickConversions,viralExternalWebsitePostViewConversions,viralFollows,viralFullScreenPlays,viralImpressions,viralJobApplications,viralJobApplyClicks,viralLandingPageClicks,viralLikes,dateRange"
+  }, {
+    "campaign_id": 123,
+    "dateRange": {
+      "end.day": 2,
+      "end.month": 3,
+      "end.year": 2021,
+      "start.day": 31,
+      "start.month": 1,
+      "start.year": 2021
+    },
+    "fields": "viralOneClickLeadFormOpens,viralOneClickLeads,viralOtherEngagements,viralPostClickJobApplications,viralPostClickJobApplyClicks,viralPostClickRegistrations,viralPostViewJobApplications,viralPostViewJobApplyClicks,viralPostViewRegistrations,viralReactions,viralRegistrations,viralShares,viralTotalEngagements,viralVideoCompletions,viralVideoFirstQuartileCompletions,viralVideoMidpointCompletions,viralVideoStarts,viralVideoThirdQuartileCompletions,viralVideoViews,dateRange"
+  }
+  ]
+]
